Move decryption function out of apple-specific package
